added responses to data dict (#19) and refactored data pipeline to use Dict[Dict] instead of List[Dict]
